Some changes
Removes start and end time from exam data handling

Eliminates the handling of start and end times in exam data as part of the schema simplification.

Updates the client dashboard to reflect this change, removing unnecessary UI elements related to start times.

This streamlines the exam data model and improves the user interface.